Revert commit to ignore VCS dotfiles
The previous commit to ignore VCS dotfiles did not work.
The code that was modified was to select which vcs command to run,
not to exclude files. The result was a vcs was not selected,
and therefore files under the control of the vcs were not always
included.

To ignore files, a future change is needed to either modify the flist
variable, or use a vcs specific exclude flag, such one of
* git -x
* git -X
* git --exclude-standard