Ensure there is a file to run checksums on
If you make a zip package (for example) of a devel build
and then try to make checksums for that package,
(using smith build; smith zip; smith checksums)
the checksum command will hang as devel package builds are
excluded from checksum calculations. With no other package files
to run on, checksum will hang, waiting for input.

This issue also occurred in some cases even if making a release,
if the package name had '-dev' in it (for example '-deva')
then the checksums command would also hang.