Fix: command operating on wrong version
Somewhere along the line a `cd` command in the namespace set up got removed. This is necessary as without it the user supplied command will operate on the original directory `$CDIR` is mounted over, and not in `$CDIR` as intended. The subsequent `rsync`, intended to make sure those changes will be propagated into the callers current dir will then end up removing them as the contents or `$CDIR` will not have changed.