Docker: Add entrypoint to setup the builder user
For interactive mode add a means to specify which UID the user inside
the container should have at runtime.  This allows the `builder` user to
access and create files as the user that launched the container's
interactive shell avoiding any ownership or permission problems, when the host user and container user don't line up exactly.