Docker: force use of legacy pip resolver
The new pip resolver seems to have a bug where it reports e.g.:
`filename has 'fonttools', but metadata has 'fonttools'`
whenever extras notation is used with github sourced packages.  This forces the use of the older resovler which works as expected.
This is not a problem with packages sourced from pypi.