Docker: satisfy freetype-py and numpy from distro
We want the python package using the same libfreetype6.so as the rest of
the tools, and as this currently provided by the distro we should use the
version of freetype-py provided by the distro too.

Use the distro build of numpy as it's not on the list of bleeding edge
packages, is mature, and takes a long time to download through pypi.