Docker: User add-apt-repository after all
It's become apparent the previous solution for getting PPA lists and
keys doesn't scale and is harder to maintain.  Install and call
`add-apt-repository` into the `ppa` stage, then copy the whole
`/etc/apt` dir. This works better for the next Ubuntu release, and
removes the need for a `codename` build arg too, and removes the need to
maintain sources.lists.d in the docker context.