Docker: Fix auto uid/gid detection to avoid root
If the detected UID & GID of the /smith volume are for the root user and
group replace them with 1000 and 1000 to create an unprivileged user
inside the contianer.  This prevents creation or modification of root
owned files on a linux Docker install and ensure a builder user on MacOS
docker installs with it's UID squash remapping.