Fix missing ldconfig for build-agent target
One (or more) of the user facing package installs in the `interactive`
target runs `ldconfig` implicitly after libs are copied into
`/usr/local/lib`, but this didn't happen for target `build-agent`. This
meant that whle the libs were present the would never be used.
Add an explicit `ldconfig` invocation to the `build-agent` target.