Only put a file into the tarball once
The creation of a zip file already had a check like this commit
has to avoid putting a file into the archive (zip or tarball) once.
The tar format can handle the duplicate file name,
the unpacked directory does not have any duplicates.
The only issue was listing the contents of the tarball
showed the duplicates, which made verification of packaging
harder.